Using hooks together in content manager
